Improves & creates tests for splitStreetAndNumber
Also renames test file from gm.test.ts to address.test.ts
