--- packages/server/src/service/gm.test.ts
@@ -1,18 +1,0 @@
-import { formatUnit } from '../common'
-
-describe('formatUnit removes unnecessary pounds', () => {
-  const unit = [
-    { raw: 'BLDG # 3', expected: 'BLDG 3' },
-    { raw: 'APT #3', expected: 'APT 3' },
-    { raw: 'LOT 3', expected: 'LOT 3' }, // should do nothing to correct ones
-    { raw: '3', expected: '# 3' },
-    { raw: '#3', expected: '# 3' },
-  ]
-
-  test.each(unit)(
-    'formatUnit works for %s',
-    input => {
-      expect(formatUnit(input.raw)).toBe(input.expected)
-    },
-  )
-})
